Updated to latest gems. Added self-contained acceptance test and removed broken integration tests

+++ spec/support/vm_control.rb
@@ -14,6 +14,7 @@ class VmControl
   end
 
   def kill
+    return unless state =~ /running/
     vboxmanage "controlvm #{@vm} poweroff"
   end
